# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cadife Smart Travel API",
    "description": "Backend inteligente para turismo via WhatsApp + Flutter. Orquestra webhooks da Meta, processamento de IA (RAG + LangChain), gestão de leads, propostas e agendamentos. Documentação completa disponível em /docs (Swagger UI) e /redoc (ReDoc).",
    "contact": {
      "name": "Cadife Tour - Time de Desenvolvimento",
      "url": "https://cadifetour.com.br/",
      "email": "dev@cadifetour.com.br"
    },
    "license": {
      "name": "Confidencial — Uso Interno do Time de Desenvolvimento"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Listar consultores",
        "description": "Lista todos os consultores com status (ativo/inativo) e métricas resumidas.",
        "operationId": "list_users_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Criar consultor",
        "description": "Cria um novo consultor com senha temporária e envia e-mail de boas-vindas.",
        "operationId": "create_user_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "E-mail já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/users/{user_id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Editar consultor",
        "description": "Edita dados do consultor ou desativa a conta (is_active = false).",
        "operationId": "update_user_admin_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Consultor não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Desativar consultor (soft delete)",
        "description": "Desativa a conta do consultor. Pode reatribuir leads ativos para outro consultor via query param.",
        "operationId": "delete_user_admin_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "reassign_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID do consultor para reatribuir leads ativos",
              "title": "Reassign To"
            },
            "description": "ID do consultor para reatribuir leads ativos"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Consultor de destino inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Consultor não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/leads/{lead_id}/reassign": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Reatribuir lead",
        "description": "Reatribui um lead para outro consultor com notificação push para ambos.",
        "operationId": "reassign_lead_admin_leads__lead_id__reassign_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLeadReassignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLeadReassignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação (mesmo consultor ou destino inválido)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Lead não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminLeadReassignRequest": {
        "properties": {
          "new_consultor_id": {
            "type": "string",
            "format": "uuid",
            "title": "New Consultor Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "new_consultor_id"
        ],
        "title": "AdminLeadReassignRequest",
        "description": "Payload for reassigning a lead to another consultant."
      },
      "AdminLeadReassignResponse": {
        "properties": {
          "lead_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lead Id"
          },
          "old_consultor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Consultor Id"
          },
          "new_consultor_id": {
            "type": "string",
            "format": "uuid",
            "title": "New Consultor Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Lead reatribuído com sucesso"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lead_id",
          "old_consultor_id",
          "new_consultor_id"
        ],
        "title": "AdminLeadReassignResponse",
        "description": "Response after reassigning a lead."
      },
      "AdminUserCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "role": {
            "$ref": "#/components/schemas/UserPerfil",
            "default": "consultor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "email"
        ],
        "title": "AdminUserCreate",
        "description": "Payload for creating a new consultant by admin."
      },
      "AdminUserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AdminUserListResponse",
        "description": "Paginated list of consultants."
      },
      "AdminUserMetrics": {
        "properties": {
          "total_leads": {
            "type": "integer",
            "title": "Total Leads",
            "default": 0
          },
          "active_leads": {
            "type": "integer",
            "title": "Active Leads",
            "default": 0
          },
          "closed_leads": {
            "type": "integer",
            "title": "Closed Leads",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdminUserMetrics",
        "description": "Aggregated metrics for a consultant."
      },
      "AdminUserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "perfil": {
            "$ref": "#/components/schemas/UserPerfil"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "metrics": {
            "$ref": "#/components/schemas/AdminUserMetrics",
            "default": {
              "total_leads": 0,
              "active_leads": 0,
              "closed_leads": 0
            }
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email",
          "telefone",
          "perfil",
          "is_active",
          "criado_em"
        ],
        "title": "AdminUserResponse",
        "description": "Consultant detail returned by admin endpoints."
      },
      "AdminUserUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdminUserUpdate",
        "description": "Payload for updating a consultant by admin."
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPErrorResponse",
        "description": "Generic error envelope returned by all failed requests.",
        "example": {
          "detail": "Descrição legível do erro",
          "error_code": "ERR_EXAMPLE"
        }
      },
      "HTTPValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPValidationErrorResponse",
        "description": "Pydantic / FastAPI validation error (422).",
        "example": {
          "detail": [
            {
              "loc": [
                "body",
                "telefone"
              ],
              "msg": "field required",
              "type": "missing"
            }
          ]
        }
      },
...
      "UserPerfil": {
        "type": "string",
        "enum": [
          "agencia",
          "cliente",
          "consultor",
          "admin"
        ],
        "title": "UserPerfil"
      },
...
      "ValidationErrorItem": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationErrorItem"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}